pyproject.toml
[project]
name = "revealpy"
version = "0.1.0"
description = "RevealPy is a Python library for creating beautiful and interactive presentations programmatically. It provides a simple yet powerful API to create presentations with various content types and layouts"
authors = [
    {name = "Davysson Santana",email = "davyssonlucas@gmail.com"}
]
readme = "README.md"
homepage = "https://github.com/davyssonlucas/revealpy.git"
repository = "https://github.com/davyssonlucas/revealpy.git"
license = "MIT"
requires-python = ">=3.12"
dependencies = [
    "python-pptx (>=1.0.2,<2.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "cairosvg (>=2.7.1,<3.0.0)"
]
keywords = ["presentations", "python-pptx", "slides", "interactive presentations"]
classifiers = [
   "Development Status :: 3 - Alpha",
   "Programming Language :: Python :: 3.12",
   "License :: OSI Approved :: MIT License",
   "Intended Audience :: Developers",
   "Topic :: Software Development :: Libraries :: Python Modules"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
setuptools = "^75.8.0"
wheel = "^0.45.1"
twine = "^6.1.0"

